Home

SRK.jl

Stochastic Runge-Kutta methods in Julia

This package tries to implement state-of-the-art SRK methods for solving stochastic differential equations (SDEs).

Functions

SRK.simdoubleintegrals — Method.

simdoubleintegrals(W::Vector{Float64}, n::Integer)

Simulates an approximation of all one-time iterated Itô-integrals of the given Brownian motions with step size 1. The algorithm is taken from [Wiktorsson2001].

Input: W the increments of m Brownian motions, where m = length(W) n number of terms in the approximation of the stochastic area integral Output: I[i,j] is an approximation of $\int_0^hW_i(s)dW_j(s)$

References

[Wiktorsson2001]
"Joint characteristic function and simultaneous simulation of iterated Itô integrals for multiple independent Brownian motions." The Annals of Applied Probability 11.2: 470-487.

source

Index

SRK.simdoubleintegrals