Home

SRK.jl

Stochastic Runge-Kutta methods in Julia

This package tries to implement state-of-the-art SRK methods for solving stochastic differential equations (SDEs).

Exported Functions

SRK.em — Function.

em(drift::Function, diffusion::Function,
    𝓘::AbstractRange,
    m::Integer, X₀::AbstractVector{<:Real},
    dW::Union{AbstractArray,Nothing}=nothing)

TODO: write docs

source

SRK.simdoubleintegrals — Function.

simdoubleintegrals(W::AbstractVector{AbstractFloat64}, h::Real, C::Real=1.0)
simdoubleintegrals(W::Real, h::Real=1.0)

Simulates an approximation of the iterated Itô-integrals $\int_0^h\int_0^sdW_i(t)dW_j(s)$ for all pairs $1\le i, j \le m$ of the given m-dimensional Brownian motion with step size h. For the used algorithm see [Wiktorsson2001].

In the case of a scalar Brownian motion the integral can be explicitly calculated as $\int_0^h\int_0^sdW(t)dW(s) = \frac{1}{2}W(h)^2 - \frac{1}{2}h$.

source

SRK.sri — Function.

sri(drift::Function, diffusion::Function,
    𝓘::AbstractRange,
    m::Integer, X₀::AbstractVector{<:Real},
    dW::Union{AbstractArray,Nothing}=nothing)

TODO: write docs

source

Non-Exported Functions

SRK.createBrownianPath — Method.

createBrownianPath(I::AbstractRange,m::Integer)

Generate an m-dimensional Brownian path at the timepoints specified by I. The path always starts at $0_m$.

source

SRK.restrict — Function.

restrict(I::AbstractRange,dW::AbstractArray,k::Integer=1)

Restrict a path of a Brownian motion given by a range of time points and the corresponding Brownian increments to $\frac{1}{2^k}$ as many time points. The lengths of I and the first dimension of dW must match and be of the form $2^n+1$ for some $n\ge k$.

source

SRK.simdoubleintegrals_n — Method.

simdoubleintegrals(W::AbstractVector{AbstractFloat64}, n::Integer)

Simulates an approximation of all one-time iterated Itô-integrals of the given Brownian motions with step size 1. The algorithm is taken from [Wiktorsson2001].

Input: W the increments of m Brownian motions, where m = length(W) n number of terms in the approximation of the stochastic area integral Output: I[i,j] is an approximation of $\int_0^1W_i(s)dW_j(s)$

source

Index

SRK.createBrownianPath

SRK.em

SRK.restrict

SRK.simdoubleintegrals

SRK.simdoubleintegrals_n

SRK.sri

References

[Wiktorsson2001]
"Joint characteristic function and simultaneous simulation of iterated Itô integrals for multiple independent Brownian motions." The Annals of Applied Probability 11.2: 470-487.